Add an "ensure" macro to check the assumption of "fileinfo"
Ensures that the fileinfo command gives the same number of infos as the status command,
which should be the number of files the Editor is enumerating.

This is not an "assert" so it won't crahs the Editor, it would trigger a breackpoint for developpers, and will log a stack in the logs to help diagnose the issue if it ever happen.