Improve performances of FPlasticCopyWorker::Execute() by not grouping the final "status" on both source & destination files
This has the potential to help a lot on big game project, where a call to status can take seconds, while it will slightly slow down the status on small project where it only takes around 100ms

Reasoning: the final "status" call on the common directory of both files could very well operate on the root Content/ folder itself, leading to an expensive call checking status of the whole project

As there is no perfect heuristic to know when a directory status is more expensive that two file status,
let's assume that a "rename" (ie. source & destination in the same directory) is always worth grouping,
while a "move" (different directories) is more at risk to create a huge status update.