Move all the Parse methods and structs from Utils to a new dedicated Parsers namespace
Typically, this moves all the code using the FXmlFile parser.

This evenly splits this big file in two and make both of them more identifiable and humanly readable.

Note: in other projects we would even split all of these into different very small files, but this isn't common practice in Unreal or in C++, as compilation time increases with the number of files