Refactor ParseFileinfoResults() to handle multiple matching locks in case of multiple destination branches
If there are multiple locks matching for a single path, concatenate all string informations (like the branch & owner)
but leave the ItmeId to invalid if there are more than one: there would be no way to know which one to unlock from the context menu.
Unlocking in such a case require using the View Locks window instead for disambiguation.