FPlasticSourceControlState::CanCheckout() now let the user try to checkout even if an asset is already checked-out elsewhere
As with multiple destination branches it is possible to have as many locks on the same asset as destination branches.

Note: Ideally we would want to check explicitly if the other lock is "on the same destination branch" as the current branch,
but that's overly complex and not possible currently with cm.