use cm shell "--enablestderr" and a dedicated Error pipe to collect error messages directly on a separate stream from the stdout regular output
Note: I added a cm version check 11.0.16.8044 (Jun 08 2023)
but we cannot use it to launch the shell as we don't have the version yet: we need the shell to get the version
Thankfully, it's not a problem to start an older shell with the new argument; unknown ones are just ignored

This is needed to detect when there is no Project accessible to the using in the @unity organization, as the CLI doesn't return an error code.
cm repository list
There are no repositories in this server.
echo %ERRORLEVEL%
0

This isn't supported by CreateProc() on UE4.27, but only from UE5.0 so it would be our very first UE5 only feature
Time to create a branch an perhaps test it one last time quickly on UE4 and update the README.md with this info

This new behavior was introduced in June 2023